error
when expense page relods the token is easily accessed from the header of the url in auth.js file but for other routes the token is undefined in auth.js file. Therefore for this reason  
when i do any other action on expense page like add expense it throw error that jwt require token.