Add files via upload
when i reload the expense page the token is accessed from headers easily but for other functionality the token which is accessed becomes undefined. I use auth.js middleware to authenticate this auth.js middleare is working perfectly for window reload but not working for other routes , throwing error that jwt require token.